Reorganize sources
File structure:
  - src/*.* contains glue code to define the extension module
  - src/scorepy contains namespaced implementation

Introduced optimization barries by using multiple cpp files can be
removed by using LinkTimeOptimization and (optionally) visibility.
